move bulk solver stuff to sub project
