+++ src/levelGenerator/solver-output.ts
@@ -1,0 +1,3 @@
+module.exports.levels = [2392, 441, 2604, 1366, 1533, 2166, 2433, 1274, 780, 1141, 2638, 2676, 363, 2929, 1246, 1913, 817, 2527, 1385, 2184, 2282, 1561, 1917, 197, 2747, 2236, 2396, 1040, 2384, 2022, 314, 1041, 592, 489, 266, 1251, 527, 1394, 2992, 2463, 664, 2377, 673, 621, 2804, 799, 253, 1863, 648, 2354, 2434, 76, 1112, 473, 1229, 742, 218, 804, 2098, 829, 844, 2591, 957, 943, 2471, 1555, 1132, 2942, 1864, 454, 947, 769, 2108, 1288, 2089, 615, 2168, 2152, 876, 867, 2345, 1784, 2385, 434, 1786, 696, 1869, 2047, 2984, 1706, 2972, 1919, 907, 1252, 2798, 733, 1631, 1809, 425, 19, 2158, 2968, 508, 1625, 2006, 702, 1710, 2550, 538, 163, 1875, 1549, 598, 204, 2693, 1815, 2044, 45, 1284, 265, 1529, 1169, 2853, 2797, 1975, 659, 537, 1501, 68, 2979, 357, 2870, 109, 2645, 140, 344, 132, 2667, 2669, 1193, 1479, 252, 135, 2762, 986, 2937, 763, 1398, 41, 60, 755, 954, 865, 1534, 2356, 1845, 774, 249, 679, 1105, 915, 645, 658, 21, 600, 911, 2083, 1497, 349, 2504, 2657, 2148, 2424, 175, 631, 1047, 377, 1867, 1500, 2617, 2592, 1254, 1382, 857, 141, 426, 2596, 879, 6, 1302, 1103, 1518, 1694, 1042, 597, 2670, 2872, 1648, 2903, 2193, 2723, 1128, 2404, 56, 1626, 545, 224, 1900, 1570, 2894, 2709, 2085, 2335, 2952, 1326, 2687, 748, 2352, 2860, 518, 1350, 1689, 2413, 926, 920, 198, 1959, 78, 365, 483, 1954, 1808, 1950, 1153, 1902, 781, 1309, 24, 521, 2128, 2037, 2318, 670, 2823, 2635, 1611, 1019, 1796, 1552, 2330, 1014, 783, 2450, 1793, 1674, 230, 113, 1410, 2386, 650, 2397, 355, 1741, 1665, 291, 2568, 812, 2363, 2362, 406, 1380, 1526, 1260, 593, 2407, 1012, 2695, 720, 211, 2739, 2859, 1026, 1732, 2456, 2801, 319, 401, 1320, 1764, 913, 889, 471, 474, 383, 1587, 149, 2817, 1754, 1458, 1712, 1139, 2220, 303, 2155, 2302, 1717, 622, 2960, 2559, 2501, 2264, 979, 206, 1573, 2176, 2322, 1211, 2833, 1220, 1806, 2014, 1206, 2503, 1769, 751, 1361, 372, 1800, 326, 369, 2321, 488, 2773, 2996, 1215, 1895, 1619, 654, 1600, 1245, 1916, 1414, 2068, 902, 603, 1126, 2628, 1761, 2112, 53, 1914, 125, 348, 2582, 638, 1007, 909, 2366, 2512, 2877, 1779, 1871, 516, 1989, 2378, 2131, 302, 1142, 2593, 1788, 2454, 51, 725, 72, 794, 1476, 2408, 2095, 2118, 614, 96, 1277, 1608, 681, 1577, 1444, 345, 2024, 2810, 1197, 2410, 1812, 707, 1690, 1456, 378, 2514, 294, 2081, 2069, 1767, 779, 2849, 2124, 165, 2115, 2867, 2696, 1267, 1948, 2460, 2051, 2651, 178, 2147, 520, 1205, 932, 2646, 394, 248, 2452, 1987, 2544, 1411, 595, 1150, 1078, 220, 2531, 1381, 2082, 2256, 1436, 1087, 1018, 77, 2125, 504, 1657, 350, 1759, 342, 2985, 1709, 1671, 566, 1313, 2704, 807, 2953, 17, 57, 1598, 2535, 2729, 2307, 458, 2700, 2324, 2547, 2683, 2852, 2041, 1675, 1152, 1890, 1581, 2389, 2671, 257, 2136, 1661, 1557, 1192, 610, 20, 2735, 632, 301, 2094, 1545, 1327, 49, 1760, 91, 616, 1866, 1655, 1264, 1262, 439, 1925, 1237, 1184, 2586, 2143, 442, 2691, 680, 1589, 2088, 2898, 1883, 114, 323, 1956, 444, 288, 1765, 389, 228, 2009, 2021, 2353, 585, 2831, 2970, 543, 2059, 2238, 195, 482, 1775, 167, 1607, 2777, 2057, 147, 2725, 13, 1530, 2430, 1445, 1704, 1544, 2913, 1196, 1332, 158, 1278, 1990, 1076, 782, 2840, 2478, 1420, 2241, 519, 1469, 2004, 858, 368, 1575, 43, 1005, 1269, 485, 156, 1490, 477, 917, 864, 1610, 2455, 2611, 624, 2949, 2177, 1707, 1967, 2694, 576, 2419, 1507, 719, 38, 1071, 530, 2313, 1470, 1715, 418, 1433, 1664, 466, 2887, 2607, 470, 1756, 1841, 1853, 2050, 2237, 82, 802, 2986, 1149, 2948, 164, 2724, 481, 1838, 2314, 724, 1749, 1351, 388, 1884, 2957, 2761, 560, 935, 1249, 1059, 107, 2150, 653, 2086, 27, 1762, 2399, 644, 493, 129, 1051, 2228, 1560, 544, 1737, 1191, 430, 300, 1316, 2371, 2260, 31, 1189, 835, 928, 2789, 2198, 1325, 1912, 296, 1918, 1783, 174, 1725, 1683, 701, 2962, 2807, 1568, 937, 1794, 2975, 2594, 1847, 1499, 2393, 586, 2780, 67, 1742, 2621, 883, 2606, 2007, 1064, 447, 1532, 2861, 2173, 801, 1880, 486, 200, 512, 1602, 81, 2064, 64, 1368, 1928, 241, 450, 851, 325, 1089, 1477, 2032, 960, 1111, 1077, 1404, 2428, 988, 2549, 1218, 2767, 318, 1407, 1542, 2449, 1113, 582, 2605, 1091, 1073, 974, 1423, 1298, 1906, 2603, 437, 1736, 2663, 1182, 1165, 964, 1836, 361, 1261, 623, 1453, 1175, 1020, 1508, 2288, 448, 1352, 1285, 2545, 709, 2253, 14, 787, 1273, 2372, 370, 313, 2388, 499, 2250, 2326, 2349, 92, 1395, 1403, 153, 941, 1744, 2634, 1688, 2980, 26, 2197, 672, 203, 54, 613, 873, 2278, 208, 1170, 182, 1993, 2060, 2636, 157, 2981, 1347, 1498, 1015, 2763, 2164, 686, 1143, 1595, 166, 2126, 66, 101, 1932, 2863, 550, 1365, 2151, 2344, 1330, 1122, 2279, 452, 2341, 1294, 152, 1094, 561, 1360, 1045, 1010, 1872, 1430, 367, 254, 251, 2997, 989, 2920, 1058, 478, 700, 517, 1198, 2141, 2521, 1495, 912, 2958, 2325, 1830, 1386, 1601, 816, 1770, 1837, 1971, 1548, 1021, 2030, 1118, 2959, 848, 1061, 2411, 2750, 1991, 2137, 124, 1405, 637, 2405, 2402, 2291, 690, 1627, 1639, 2205, 460, 599, 1310, 1701, 2329, 34, 1915, 2644, 212, 602, 2755, 2199, 1816, 1851, 756, 2285, 2226, 2267, 546, 892, 959, 1367, 860, 2145, 1716, 930, 2779, 243, 1209, 59, 1371, 2885, 1659, 1217, 1257, 503, 590, 1355, 188, 572, 1339, 973, 2637, 1811, 459, 983, 2677, 2310, 384, 142, 2581, 456, 1208, 2342, 1074, 1409, 2934, 2791, 1992, 1119, 333, 906, 1253, 1315, 39, 1455, 646, 2170, 1099, 1949, 131, 2619, 790, 2172, 2924, 1714, 2757, 1844, 419, 2270, 2778, 940, 2074, 2631, 1465, 2630, 2941, 1592, 2871, 2911, 2380, 2053, 2546, 2982, 1825, 2744, 1538, 2360, 2865, 237, 1234, 276, 2276, 1700, 1321, 1571, 199, 117, 2422, 2976, 747, 661, 640, 2300, 1795, 677, 806, 1556, 2465, 2489, 2699, 832, 577, 127, 1024, 862, 1832, 1081, 1818, 715, 1226, 1617, 1968, 2135, 305, 2028, 958, 382, 1334, 159, 2080, 415, 2844, 395, 1774, 2162, 2771, 1964, 2523, 122, 2850, 1780, 1998, 1194, 2579, 2846, 1109, 416, 1624, 975, 868, 209, 2918, 1996, 1748, 2584, 2991, 1060, 1565, 2673, 1349, 1613, 1974, 2775, 2099, 990, 1039, 1137, 541, 1947, 231, 2469, 1319, 258, 2907, 2742, 2814, 1672, 1317, 1450, 2289, 2746, 2446, 2251, 788, 2774, 2540, 2556, 1614, 1364, 2654, 606, 967, 2346, 662, 1480, 2340, 1413, 2543, 42, 106, 280, 811, 971, 2749, 1243, 2510, 854, 337, 761, 2283, 358, 1457, 2048, 2365, 2827, 2908, 1066, 233, 1304, 2845, 910, 1108, 2803, 924, 1029, 2156, 2105, 1402, 1408, 308, 1585, 760, 1124, 1106, 1145, 1155, 994, 145, 2394, 2298, 2821, 2192, 263, 1286, 2674, 1509, 2825, 2608, 1910, 2311, 2808, 1107, 2447, 1908, 944, 79, 148, 2225, 317, 2945, 651, 1642, 931, 895, 2928, 80, 569, 2625, 934, 1797, 555, 861, 1161, 815, 901, 2541, 2333, 657, 805, 2448, 691, 1758, 505, 1713, 1095, 1590, 247, 2662, 1167, 2400, 137, 330, 1696, 1296, 126, 371, 315, 1426, 568, 356, 2093, 1162, 2989, 118, 134, 562, 617, 1773, 2211, 2765, 625, 501, 642, 1941, 1322, 2071, 1300, 1905, 95, 2106, 2142, 2711, 2001, 2886, 683, 2525, 1297, 1006, 2458, 866, 674, 2063, 2210, 1336, 162, 2506, 1116, 852, 2697, 1397, 1013, 1596, 340, 284, 953, 1735, 1541, 2293, 480, 2839, 2665, 1125, 1576, 1163, 881, 2065, 52, 754, 1582, 753, 1305, 2107, 1179, 2805, 1129, 1746, 2049, 2373, 921, 2713, 1678, 1067, 2167, 277, 968, 1646, 1337, 1554, 2429, 1846, 381, 85, 1649, 2257, 2056, 1238, 1645, 1831, 738, 223, 2488, 1362, 1159, 1399, 2383, 1966, 2387, 796, 2367, 2562, 540, 1008, 2588, 511, 2415, 2248, 271, 652, 1030, 268, 2027, 334, 1097, 1307, 1486, 1663, 1891, 1922, 2436, 2438, 2862, 2946, 1471, 311, 570, 1461, 2390, 221, 1512, 918, 1708, 1584, 99, 1594, 1985, 2902, 2919, 2933, 1876, 605, 2685, 2795, 2505, 708, 1390, 411, 1492, 1003, 1896, 1070, 890, 2266, 2327, 1750, 878, 1199, 2553, 1798, 2879, 735, 750, 16, 2087, 1358, 1721, 1329, 1583, 1188, 2133, 1660, 559, 1877, 539, 1942, 207, 2453, 2374, 1046, 2737, 1753, 710, 2818, 12, 2026, 279, 564, 449, 2181, 177, 2802, 813, 40, 587, 639, 2018, 1176, 1647, 1502, 2437, 942, 1591, 739, 1416, 847, 1731, 982, 778, 728, 1698, 1102, 992, 2286, 1817, 2000, 63, 328, 392, 413, 509, 803, 818, 822, 896, 1281, 1640, 1751, 2284, 2599, 1889, 2010, 2782, 409, 1090, 1703, 1244, 2794, 1979, 2639, 1166, 2585, 2719, 939, 2169, 1506, 1120, 2391, 2643, 1428, 387, 2008, 1505, 2140, 1389, 2435, 916, 771, 2308, 2336, 2224, 2668, 1637, 428, 2468, 1807, 36, 1632, 1177, 35, 1121, 183, 1065, 2738, 1442, 1835, 281, 588, 2062, 1447, 579, 2743, 2163, 1976, 893, 2425, 1424, 298, 1472, 267, 2440, 2832, 1702, 1181, 1473, 2232, 407, 843, 2243, 44, 75, 2716, 970, 1485, 1927, 312, 2653, 2406, 946, 636, 307, 1562, 2412, 950, 604, 11, 86, 777, 877, 1384, 1790, 1888, 2090, 2678, 2705, 2720, 1117, 1233, 351, 170, 1235, 30, 2703, 2589, 58, 2092, 1588, 2149, 123, 2245, 1638, 494, 2078, 2482, 2515, 343, 1658, 309, 2091, 1995, 694, 2188, 2909, 2983, 1960, 1650, 563, 119, 2111, 875, 297, 1210, 1036, 2480, 1734, 2943, 1958, 2139, 2642, 594, 1048, 1096, 202, 2013, 1222, 999, 2320, 2712, 1874, 1599, 1718, 1878, 904, 1088, 2120, 1819, 1265, 1782, 1443, 171, 2339, 900, 2536, 304, 1270, 393, 429, 427, 744, 475, 704, 1331, 1564, 2401, 2600, 2822, 2921, 1275, 1158, 609, 831, 2904, 1353, 2978, 2753, 2160, 923, 2323, 955, 2132, 2464, 1898, 93, 1348, 2475, 1068, 1911, 1001, 97, 1190, 687, 2973, 1022, 196, 1618, 354, 634, 1482, 618, 2096, 2215, 128, 2598, 2577, 608, 849, 1517, 1291, 1951, 2679, 2689, 105, 2938, 2239, 2680, 703, 1666, 767, 232, 908, 2754, 2343, 1822, 2566, 10, 273, 1216, 1682, 2935, 2717, 2681, 1038, 897, 1662, 1266, 1550, 2262, 172, 1114, 2025, 2035, 2017, 1711, 497, 1622, 1687, 1930, 726, 2613, 2208, 1728, 2561, 1263, 717, 2848, 1909, 2212, 2421, 269, 2016, 1747, 2575, 353, 2868, 2787, 347, 2623, 1172, 2912, 764, 1612, 1272, 2474, 905, 663, 1805, 1559, 1359, 1858, 978, 2121, 762, 1551, 88, 1894, 2655, 422, 1392, 1729, 1523, 2652, 629, 168, 2570, 1514, 403, 2055, 1053, 1824, 1567, 69, 1776, 1978, 1055, 1605, 1224, 321, 2491, 2792, 2878, 225, 827, 2590, 1415, 2466, 1952, 446, 2627, 2219, 71, 120, 133, 689, 1050, 1848, 1850, 2610, 1536, 2203, 2265, 2054, 1454, 2222, 1201, 1123, 1961, 1136, 529, 552, 2830, 2242, 532, 1572, 1861, 1723, 1011, 2740, 2641, 1363, 1969, 1100, 1724, 116, 1686, 837, 1516, 2650, 2296, 850, 1451, 1524, 1651, 2235, 2571, 1903, 743, 506, 514, 2294, 2616, 2259, 373, 1299, 476, 386, 1419, 227, 2816, 410, 2910, 1396, 32, 2109, 666, 786, 1446, 2122, 1069, 1868, 2217, 2828, 431, 412, 2015, 976, 1154, 2955, 2509, 1412, 2207, 1586, 705, 823, 2969, 526, 919, 2261, 2922, 2736, 121, 1354, 1341, 2364, 1670, 792, 2102, 619, 1440, 2230, 685, 2066, 155, 2715, 2812, 1417, 2722, 1236, 2481, 1679, 2819, 320, 2710, 1965, 2249, 730, 1934, 2648, 1763, 1475, 2752, 285, 1699, 1314, 1945, 1467, 2843, 272, 396, 647, 980, 264, 809, 2660, 2223, 1768, 1421, 874, 2718, 435, 2858, 871, 2472, 2800, 2297, 364, 176, 1072, 1241, 716, 15, 1228, 464, 1980, 1852, 219, 547, 2073, 2708, 2690, 628, 1540, 1563, 1086, 1183, 969, 819, 649, 115, 402, 2857, 2033, 991, 2500, 2244, 914, 2726, 2891, 1520, 18, 226, 335, 1994, 2659, 627, 213, 1778, 1434, 2721, 1757, 1733, 1988, 793, 2727, 1378, 1084, 2146, 2632, 1212, 2498, 1311, 2956, 184, 2295, 1859, 2114, 2565, 2216, 826, 2560, 201, 1075, 1521, 2395, 2485, 956, 1634, 2601, 2416, 2788, 2268, 2783, 1448, 1130, 2882, 2258, 1739, 2614, 1827, 289, 533, 238, 741, 216, 998, 1681, 2191, 1346, 1929, 1644, 2529, 2097, 1369, 2583, 1459, 1481, 2824, 1633, 469, 2067, 104, 2649, 2786, 2612, 2183, 1185, 433, 462, 2950, 2306, 1377, 1255, 1393, 286, 1885, 2334, 2351, 2834, 558, 2897, 2796, 1630, 1437, 90, 492, 1977, 2993, 2707, 1970, 89, 1820, 2609, 2075, 2031, 2629, 1899, 1148, 1202, 2252, 853, 2444, 578, 515, 214, 2965, 2039, 2273, 2467, 1338, 1653, 1343, 776, 1513, 1558, 2940, 1522, 1924, 1813, 2499, 641, 1907, 2459, 757, 549, 669, 2618, 1511, 1242, 723, 346, 1171, 2277, 138, 2043, 194, 2563, 1727, 2854, 62, 1804, 2218, 1287, 535, 405, 607, 2587, 1478, 259, 2947, 2443, 2290, 668, 1692, 840, 1752, 1056, 2555, 154, 2246, 1636, 239, 1740, 2864, 2790, 945, 400, 2672, 2005, 250, 2895, 2876, 882, 2994, 189, 2530, 1160, 1946, 612, 1418, 1435, 2728, 567, 2538, 1893, 2963, 50, 2473, 745, 1574, 1677, 2815, 130, 1983, 1009, 1035, 789, 2706, 487, 721, 2927, 179, 23, 2194, 1328, 1214, 2281, 70, 2998, 536, 1391, 414, 2263, 160, 678, 2233, 2622, 675, 1002, 399, 2200, 359, 1028, 2490, 1840, 2202, 310, 65, 2552, 810, 2513, 2906, 1628, 2533, 151, 531, 451, 1781, 1578, 2785, 1705, 846, 1643, 665, 2701, 557, 2893, 1685, 1510, 2926, 1799, 768, 1466, 190, 245, 1593, 2534, 2522, 1062, 2526, 1503, 2350, 1904, 2971, 1789, 180, 1280, 1803, 2511, 73, 108, 2961, 870, 2255, 331, 1376, 1140, 795, 841, 2019, 37, 1491, 2574, 1079, 270, 2240, 1324, 785, 33, 770, 25, 2524, 2275, 1248, 1054, 2476, 215, 1537, 2988, 329, 1422, 2119, 740, 2917, 886, 2842, 1438, 2486, 722, 421, 1494, 2951, 2572, 1085, 2516, 729, 1484, 2190];
+
+module.exports.scores = [2.307103842, 2.441768062, 2.708754955, 2.710587234, 2.767904035, 2.790480282, 2.799377566, 2.801284199, 2.811164907, 2.847450219, 2.86436751, 2.865926823, 2.929112245, 2.955378257, 3.034290432, 3.05608201, 3.058411385, 3.067494157, 3.126301815, 3.13974359, 3.195108313, 3.212480393, 3.240100765, 3.251463438, 3.280773293, 3.289593072, 3.290010111, 3.307767781, 3.343652706, 3.349232582, 3.356733347, 3.400054307, 3.40722692, 3.421044639, 3.424563999, 3.438323051, 3.440474153, 3.445142704, 3.462521826, 3.468610796, 3.469465119, 3.479712312, 3.516195206, 3.542545618, 3.564031816, 3.567382106, 3.567891662, 3.570909237, 3.582044193, 3.593335014, 3.601330744, 3.614090104, 3.614341445, 3.643862423, 3.651266202, 3.665924246, 3.674374957, 3.674535815, 3.676269506, 3.679427967, 3.683352753, 3.683629418, 3.693166212, 3.693274482, 3.693505319, 3.696324414, 3.72367613, 3.734080441, 3.741806902, 3.748193098, 3.76231239, 3.770536236, 3.771154362, 3.775153143, 3.796171414, 3.796422157, 3.800664054, 3.806665979, 3.812330721, 3.816401816, 3.831136022, 3.831550224, 3.833761686, 3.854075869, 3.858340379, 3.859237682, 3.861530513, 3.865380963, 3.887922939, 3.895999175, 3.902226527, 3.902377024, 3.903005989, 3.907517186, 3.90848705, 3.911764453, 3.92005117, 3.921134083, 3.922828762, 3.933566772, 3.938928989, 3.940515012, 3.965905966, 3.966242181, 3.968688929, 3.975683131, 3.976820607, 3.977124395, 3.982850989, 3.982854999, 3.983534215, 3.989081254, 3.992169716, 3.992610447, 3.993276964, 3.996305817, 4.023259091, 4.024577232, 4.038183131, 4.043134538, 4.052946388, 4.055618512, 4.055857307, 4.062862274, 4.065087647, 4.069655084, 4.073291314, 4.074372078, 4.074492163, 4.081919812, 4.091900821, 4.093878845, 4.09410264, 4.100395461, 4.102409241, 4.111073374, 4.115335923, 4.127599811, 4.135126893, 4.13578444, 4.140459087, 4.140788049, 4.141572145, 4.146228174, 4.148047707, 4.151848606, 4.158568041, 4.159589606, 4.16050312, 4.163805252, 4.170408479, 4.171055029, 4.172576304, 4.174034165, 4.175094521, 4.179413854, 4.192845849, 4.194343121, 4.195749983, 4.200786858, 4.201037155, 4.201290644, 4.204110384, 4.205316173, 4.208181854, 4.213215047, 4.22141872, 4.228382355, 4.229291017, 4.236954109, 4.240158409, 4.240517709, 4.244585531, 4.253396963, 4.254100717, 4.259467015, 4.260052146, 4.2626818, 4.264542517, 4.26691414, 4.274688694, 4.278908022, 4.279608338, 4.282817381, 4.284014868, 4.284214099, 4.29035319, 4.293296441, 4.29657023, 4.296628858, 4.301409798, 4.306157627, 4.314849711, 4.320229502, 4.325353834, 4.327066455, 4.33140892, 4.332239833, 4.332992484, 4.334169321, 4.336345295, 4.338340379, 4.344695355, 4.365294677, 4.366929321, 4.367485196, 4.373798292, 4.37763751, 4.385549903, 4.386117413, 4.390290953, 4.40041131, 4.401580438, 4.403178662, 4.405308655, 4.408592264, 4.409668415, 4.410158237, 4.410255981, 4.410517387, 4.412855457, 4.414128915, 4.417051912, 4.418732385, 4.420906086, 4.422467077, 4.432898879, 4.435069301, 4.441901036, 4.443059197, 4.445800203, 4.446181343, 4.44859868, 4.449592127, 4.449723311, 4.453622525, 4.455445895, 4.457573039, 4.459004262, 4.460613136, 4.464268283, 4.465170776, 4.467714993, 4.468075155, 4.46816762, 4.472544167, 4.477748677, 4.482978449, 4.483722589, 4.484346061, 4.485443161, 4.489742253, 4.49078053, 4.496370195, 4.496401535, 4.498758621, 4.500067761, 4.500977292, 4.503556698, 4.507955003, 4.517678044, 4.5196617, 4.520935343, 4.523735994, 4.530480855, 4.534890785, 4.538004154, 4.538278879, 4.539203387, 4.542845113, 4.543229131, 4.544278889, 4.546450299, 4.557118306, 4.558222915, 4.560655118, 4.563440916, 4.563669141, 4.566878296, 4.569520348, 4.571952407, 4.577746486, 4.582992484, 4.58621327, 4.590841018, 4.590991556, 4.591069464, 4.59213469, 4.594473087, 4.595325251, 4.596819275, 4.600423627, 4.603304314, 4.603636202, 4.604398158, 4.607294975, 4.610035206, 4.611306209, 4.61399143, 4.618727802, 4.621054384, 4.626539836, 4.628802545, 4.629155878, 4.630168076, 4.632070186, 4.634843095, 4.640335636, 4.643087922, 4.647591857, 4.648461023, 4.649507058, 4.6504643, 4.652009292, 4.658209842, 4.658656138, 4.65918697, 4.66301643, 4.663491982, 4.667589881, 4.668914553, 4.669804198, 4.675934901, 4.681888663, 4.682911425, 4.684881118, 4.685207799, 4.691470013, 4.692693224, 4.69624493, 4.698200015, 4.699507058, 4.700521156, 4.704078848, 4.706894892, 4.708934488, 4.713545748, 4.714396439, 4.715064962, 4.716173438, 4.719676566, 4.722106178, 4.722183554, 4.722206469, 4.72232505, 4.723470045, 4.72560021, 4.726143338, 4.729177646, 4.732943791, 4.736414438, 4.740662287, 4.742059973, 4.742149339, 4.742833574, 4.743001503, 4.743017174, 4.743148072, 4.744053542, 4.745835338, 4.74587286, 4.746738896, 4.751957939, 4.753553806, 4.757456864, 4.760782537, 4.762218728, 4.763459247, 4.766844825, 4.767395585, 4.768704151, 4.768748281, 4.769342476, 4.771152128, 4.772488887, 4.773104666, 4.774008878, 4.776465938, 4.776772273, 4.778364955, 4.778695779, 4.779153777, 4.779321911, 4.779988428, 4.784262833, 4.784742845, 4.79301832, 4.795034568, 4.79527394, 4.797011411, 4.800365711, 4.804065271, 4.804727779, 4.805293188, 4.806058441, 4.806462845, 4.807277102, 4.809384115, 4.811196896, 4.813030178, 4.813625784, 4.814168361, 4.814736715, 4.816117584, 4.816404872, 4.817727023, 4.819909985, 4.821266781, 4.82179344, 4.822644826, 4.825802857, 4.826517495, 4.826941981, 4.830890103, 4.831071126, 4.834531175, 4.834824559, 4.837431544, 4.838671204, 4.838857496, 4.839882221, 4.840773012, 4.847453427, 4.848536026, 4.850751014, 4.854842923, 4.855027153, 4.855588495, 4.855856018, 4.856275717, 4.856292535, 4.859400564, 4.859642636, 4.86451993, 4.866460439, 4.866623359, 4.867198735, 4.868071166, 4.869350496, 4.87040185, 4.870955523, 4.871722692, 4.873959694, 4.874137279, 4.875431606, 4.878425105, 4.878556207, 4.879149911, 4.884562355, 4.886749845, 4.887070358, 4.889854315, 4.889988608, 4.892663722, 4.899611031, 4.899917222, 4.902189885, 4.902280539, 4.903850082, 4.906859146, 4.911682576, 4.913559497, 4.920303155, 4.921018079, 4.921412491, 4.921546907, 4.921920041, 4.921971598, 4.922302422, 4.922377753, 4.924610744, 4.924879987, 4.925178158, 4.925830805, 4.925903966, 4.927136179, 4.929917967, 4.930505431, 4.932322816, 4.932915149, 4.933126246, 4.933610759, 4.935079398, 4.935344057, 4.935874063, 4.938155977, 4.940697108, 4.94380972, 4.947051282, 4.951870489, 4.952086914, 4.957900254, 4.958378262, 4.959878039, 4.960107754, 4.963354987, 4.966353315, 4.966565271, 4.968668798, 4.971439128, 4.97153279, 4.971805183, 4.977307005, 4.979867474, 4.980478449, 4.980588208, 4.981623359, 4.984844298, 4.985156733, 4.987718831, 4.988352581, 4.988376255, 4.990501822, 4.990717158, 4.990853784, 4.991846134, 4.993543342, 4.993720641, 4.99439833, 4.996900564, 4.998944554, 4.999566676, 5.000656063, 5.000916426, 5.00112423, 5.002198735, 5.003327146, 5.004596922, 5.00550629, 5.006876752, 5.00770176, 5.012461823, 5.015476559, 5.015547879, 5.016075823, 5.016791091, 5.019754589, 5.023145093, 5.023282578, 5.025129465, 5.025316646, 5.025590041, 5.027837698, 5.028946404, 5.02942617, 5.031224078, 5.032919159, 5.033247835, 5.038877775, 5.039528482, 5.041558741, 5.042277102, 5.043109232, 5.04404989, 5.044688596, 5.045925277, 5.045947332, 5.047245566, 5.049420597, 5.049673043, 5.051373422, 5.051745492, 5.052667904, 5.054832955, 5.057121743, 5.058339635, 5.059296762, 5.059364903, 5.060451468, 5.060843731, 5.064080223, 5.064759057, 5.066616092, 5.067534973, 5.067733324, 5.069466557, 5.071556482, 5.072194807, 5.073153829, 5.074671181, 5.075987833, 5.077091497, 5.079807177, 5.081835602, 5.082561942, 5.084714374, 5.085032996, 5.086089572, 5.086812205, 5.086832091, 5.088069018, 5.088099207, 5.088608476, 5.089429092, 5.090135423, 5.091951212, 5.092007516, 5.092588506, 5.095179075, 5.095474439, 5.098013336, 5.098814188, 5.100721798, 5.103378704, 5.107684746, 5.108127965, 5.11079604, 5.115480855, 5.119616158, 5.121953839, 5.122476799, 5.122913946, 5.126662714, 5.129319046, 5.131170087, 5.134202928, 5.136532324, 5.13703971, 5.137402328, 5.139602954, 5.140886093, 5.142239236, 5.142612094, 5.143759194, 5.144115625, 5.14494999, 5.148725682, 5.148921599, 5.149341789, 5.149908916, 5.150549369, 5.152365178, 5.153258895, 5.154459191, 5.154783461, 5.155428095, 5.157192205, 5.159512843, 5.160522215, 5.162010086, 5.16259538, 5.165461607, 5.16629653, 5.167085997, 5.167316973, 5.167898593, 5.169414901, 5.169567952, 5.170083179, 5.170132158, 5.171885465, 5.172156016, 5.173384792, 5.174016693, 5.176256273, 5.176638081, 5.17835155, 5.181293738, 5.1818462, 5.183306867, 5.18403279, 5.184830206, 5.185263972, 5.185870626, 5.188371958, 5.188781536, 5.188905502, 5.189081941, 5.190808119, 5.19404757, 5.194176062, 5.196935221, 5.197170011, 5.197632502, 5.199756536, 5.199869634, 5.20241516, 5.203788705, 5.206040077, 5.208372517, 5.208696638, 5.209836221, 5.210675741, 5.211678009, 5.213657701, 5.214878669, 5.215185605, 5.216965268, 5.218102816, 5.22140132, 5.22140132, 5.22140132, 5.223675414, 5.225555834, 5.231508215, 5.235018389, 5.238225235, 5.239375816, 5.240784354, 5.242629924, 5.245606912, 5.246132732, 5.247366582, 5.247891318, 5.247928553, 5.249984533, 5.24999198, 5.250327101, 5.250754451, 5.251453908, 5.251602134, 5.252324976, 5.252933881, 5.253625604, 5.254980981, 5.255860028, 5.257859009, 5.258244312, 5.25987317, 5.260169794, 5.265123393, 5.267738766, 5.268139479, 5.270331082, 5.270552348, 5.274231371, 5.276063793, 5.276389891, 5.276735323, 5.279262047, 5.279488898, 5.280461607, 5.281384134, 5.281411288, 5.281616239, 5.283234516, 5.285353223, 5.289637325, 5.29041154, 5.291356752, 5.292131253, 5.292996494, 5.293144635, 5.294622889, 5.297507791, 5.300213839, 5.301358814, 5.302117275, 5.302196084, 5.303331271, 5.303555716, 5.303640613, 5.304120552, 5.304691403, 5.306346669, 5.306635818, 5.306839039, 5.307944937, 5.311191884, 5.312949978, 5.314786984, 5.314801591, 5.314821613, 5.31483496, 5.316144652, 5.316177734, 5.316193774, 5.318712192, 5.319278889, 5.319927648, 5.322454802, 5.322694828, 5.323146955, 5.326199274, 5.328421883, 5.329036628, 5.331889049, 5.332993057, 5.334551798, 5.336052352, 5.337606551, 5.338522834, 5.338910085, 5.339553688, 5.341540782, 5.342122517, 5.342555166, 5.344548704, 5.346006622, 5.34827241, 5.34896336, 5.349281639, 5.351110195, 5.351268876, 5.351651543, 5.353270004, 5.35394268, 5.355678147, 5.355911872, 5.356903829, 5.359803568, 5.360563346, 5.360842442, 5.361831764, 5.362462192, 5.363483192, 5.36399544, 5.365369148, 5.366453564, 5.368793566, 5.370325497, 5.371390763, 5.372054805, 5.373266395, 5.377705713, 5.379161339, 5.379228363, 5.380665773, 5.380874636, 5.381423575, 5.381481577, 5.382419915, 5.383231938, 5.383504904, 5.388431635, 5.390622351, 5.391004331, 5.393680828, 5.394152574, 5.394348921, 5.395397185, 5.396014642, 5.400368633, 5.400547536, 5.40106322, 5.402534028, 5.404634976, 5.40666174, 5.40808987, 5.408804908, 5.410201416, 5.41112085, 5.414113612, 5.414450746, 5.423494535, 5.425656493, 5.426197841, 5.428431979, 5.428594261, 5.429734653, 5.429734653, 5.429734653, 5.429734653, 5.429734653, 5.430065993, 5.430072781, 5.430167159, 5.431384478, 5.432202086, 5.432895958, 5.436761875, 5.439035196, 5.442492152, 5.443314773, 5.443857153, 5.444545697, 5.44502131, 5.445676084, 5.446066887, 5.448174881, 5.451615453, 5.455025091, 5.455385561, 5.456232098, 5.456779748, 5.457007172, 5.459560963, 5.461590878, 5.463084313, 5.463818141, 5.464075411, 5.46491115, 5.465622923, 5.46594315, 5.466565271, 5.46692216, 5.467311559, 5.468177287, 5.470381718, 5.470529662, 5.470623668, 5.470874149, 5.472545886, 5.473286015, 5.474633945, 5.475192136, 5.476312985, 5.478521345, 5.479013886, 5.479710593, 5.482703765, 5.482930616, 5.485262597, 5.485682042, 5.490314154, 5.498107284, 5.498869526, 5.49986853, 5.500100536, 5.500625988, 5.501126807, 5.501133272, 5.501689008, 5.501758524, 5.503659403, 5.505727298, 5.506619922, 5.508419605, 5.508950299, 5.509434935, 5.509996907, 5.511061387, 5.511547398, 5.514121468, 5.514144669, 5.51444685, 5.514565718, 5.515239683, 5.515730907, 5.51783787, 5.51834628, 5.520405926, 5.520449234, 5.520779198, 5.521476937, 5.521811198, 5.521938659, 5.52548349, 5.526701382, 5.527215663, 5.528316835, 5.530253088, 5.531246993, 5.532880061, 5.533238468, 5.533797347, 5.534684987, 5.535786588, 5.536318485, 5.537524919, 5.537765948, 5.538284526, 5.539156985, 5.541569568, 5.541951949, 5.543732385, 5.545349557, 5.546818932, 5.549511354, 5.549524301, 5.549760088, 5.549920087, 5.550082921, 5.551877105, 5.553122637, 5.553122637, 5.553122637, 5.553122637, 5.553122637, 5.553122637, 5.554289544, 5.554663762, 5.555055166, 5.555315099, 5.555399309, 5.55697767, 5.557117017, 5.557855572, 5.558466236, 5.560941775, 5.561988382, 5.564964426, 5.568596045, 5.569504021, 5.571542815, 5.571948512, 5.572176394, 5.572719461, 5.572999588, 5.573054582, 5.575699457, 5.577101464, 5.577511564, 5.577808483, 5.579832612, 5.580365367, 5.580535506, 5.583183245, 5.583588827, 5.585264144, 5.585421393, 5.586994226, 5.589137795, 5.59147032, 5.593445384, 5.594562453, 5.595014436, 5.595917543, 5.599375014, 5.600218258, 5.600359753, 5.600572283, 5.603424532, 5.604818433, 5.605505001, 5.606021001, 5.606826723, 5.60788092, 5.612011068, 5.613605898, 5.618971695, 5.62065718, 5.621242352, 5.625162691, 5.626976353, 5.628290025, 5.629202757, 5.630591645, 5.630631316, 5.631847345, 5.632416478, 5.634174572, 5.635658211, 5.636407278, 5.638538874, 5.638586994, 5.639208542, 5.640112345, 5.640183543, 5.640439953, 5.64197028, 5.642685777, 5.643074603, 5.645735375, 5.646465766, 5.646956417, 5.648864027, 5.64933079, 5.650625559, 5.65241516, 5.655064045, 5.655647785, 5.656727332, 5.659650615, 5.660092803, 5.662892263, 5.663582697, 5.664573452, 5.66475562, 5.665063673, 5.665217914, 5.665339933, 5.667441741, 5.667779611, 5.668256055, 5.669090706, 5.671483811, 5.672042346, 5.673681859, 5.674314292, 5.676510621, 5.676510621, 5.676510621, 5.676510621, 5.676510621, 5.676510621, 5.678357221, 5.680178731, 5.68122551, 5.682848697, 5.68369366, 5.684312057, 5.684456704, 5.686191311, 5.686499364, 5.689398158, 5.690168076, 5.69067849, 5.690984395, 5.691699319, 5.691922733, 5.692412869, 5.6940824, 5.694491991, 5.694751066, 5.696272141, 5.696853303, 5.697899051, 5.69825531, 5.69874029, 5.698927041, 5.69902815, 5.699615041, 5.702237575, 5.703517907, 5.703875942, 5.705172888, 5.705631316, 5.706575239, 5.707108486, 5.710086272, 5.710240514, 5.712100433, 5.712719117, 5.714915446, 5.717819089, 5.718959236, 5.719545267, 5.725230861, 5.725345432, 5.726507184, 5.728965021, 5.729687393, 5.732956967, 5.734882335, 5.735275314, 5.735297656, 5.735783151, 5.738153571, 5.740239683, 5.740574517, 5.741602507, 5.7422936, 5.745673507, 5.749084433, 5.752519993, 5.753165601, 5.75317935, 5.75349316, 5.757676411, 5.762943562, 5.765645838, 5.765791057, 5.767405651, 5.769636924, 5.77113709, 5.77213996, 5.776141129, 5.776565615, 5.776849179, 5.78217373, 5.785898009, 5.78775578, 5.788727916, 5.790150317, 5.792414243, 5.792855486, 5.793797261, 5.794369114, 5.796818645, 5.798111294, 5.798484224, 5.798606242, 5.799898605, 5.799898605, 5.799898605, 5.799898605, 5.799898605, 5.799898605, 5.799898605, 5.800378658, 5.800943923, 5.801831134, 5.802132742, 5.802559291, 5.802617378, 5.804743819, 5.805274627, 5.805455592, 5.805912445, 5.810168763, 5.810523132, 5.812038565, 5.812793875, 5.813235833, 5.814651933, 5.815187553, 5.817120941, 5.817311129, 5.817961607, 5.818334708, 5.818787035, 5.82120884, 5.822939008, 5.824991407, 5.826564927, 5.827729085, 5.827754864, 5.834353017, 5.835699801, 5.835753076, 5.838274501, 5.838341239, 5.838860676, 5.840537568, 5.840984395, 5.842128961, 5.843822896, 5.845602186, 5.847748247, 5.847836324, 5.850461607, 5.850973998, 5.85231577, 5.855795353, 5.856300268, 5.857184185, 5.858155032, 5.858452201, 5.859565546, 5.860894858, 5.863607187, 5.865471117, 5.865575376, 5.866836805, 5.866959511, 5.871934076, 5.873894274, 5.873942651, 5.876958308, 5.879369286, 5.879382347, 5.879817144, 5.880205541, 5.881404241, 5.881793497, 5.882334846, 5.882900827, 5.883593696, 5.885101739, 5.885667148, 5.886465136, 5.88992318, 5.892955592, 5.897452739, 5.89815254, 5.898297759, 5.898605669, 5.899244861, 5.899558758, 5.899638242, 5.900118581, 5.90317935, 5.904710593, 5.905413487, 5.906426583, 5.906805097, 5.907219702, 5.907856717, 5.909296161, 5.910060494, 5.910206916, 5.910861919, 5.912257396, 5.915555326, 5.916748299, 5.91746666, 5.91812974, 5.919108014, 5.919354506, 5.921482522, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.923286588, 5.925520726, 5.927609645, 5.927754864, 5.927862274, 5.933467382, 5.933928817, 5.936691414, 5.94079879, 5.941670104, 5.942176508, 5.944863632, 5.94508948, 5.946511308, 5.947404104, 5.949827714, 5.949984533, 5.950472606, 5.951291503, 5.951512339, 5.953323709, 5.956552325, 5.956835888, 5.959748058, 5.959904104, 5.961729223, 5.967280768, 5.967596412, 5.967846177, 5.96899017, 5.971741596, 5.973458445, 5.974501615, 5.975062728, 5.975783667, 5.978772255, 5.980096412, 5.984523441, 5.985612154, 5.986304736, 5.986535024, 5.989044992, 5.990773012, 5.991000435, 5.991073761, 5.991130474, 5.993874453, 5.994602409, 5.996201966, 5.997042346, 5.997245136, 5.997813982, 5.997826012, 6.000721455, 6.000970991, 6.001509332, 6.002280109, 6.002666357, 6.004921977, 6.005096412, 6.006624218, 6.008183818, 6.009249903, 6.009457276, 6.01032859, 6.010332027, 6.013143603, 6.013373462, 6.016301185, 6.01909844, 6.022370592, 6.024020989, 6.02818737, 6.028424123, 6.029441924, 6.030732711, 6.030978896, 6.032147522, 6.032304599, 6.033452774, 6.034323228, 6.034904104, 6.035092717, 6.035101739, 6.038373462, 6.039072919, 6.039350038, 6.042206297, 6.042318004, 6.044778992, 6.045169966, 6.045967095, 6.046427958, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.046674572, 6.049161339, 6.050189902, 6.053008639, 6.05750613, 6.05784526, 6.058842889, 6.059457276, 6.060120643, 6.062916753, 6.063467553, 6.064372379, 6.065944296, 6.066535483, 6.069505053, 6.06975751, 6.073687874, 6.074429865, 6.075654345, 6.076359387, 6.077245595, 6.078254279, 6.079535872, 6.079813707, 6.080072781, 6.080298343, 6.081006393, 6.084653193, 6.08488921, 6.085117206, 6.086515261, 6.091922733, 6.092394996, 6.092601567, 6.09394205, 6.09593387, 6.09732694, 6.097943906, 6.098734275, 6.100138517, 6.101785162, 6.102367327, 6.102528013, 6.1035033, 6.105100708, 6.105987489, 6.106448638, 6.108385463, 6.110329449, 6.113382484, 6.114942084, 6.116024352, 6.116144222, 6.116762047, 6.11703444, 6.117385601, 6.117458926, 6.11843049, 6.119890699, 6.121626452, 6.122124837, 6.122124837, 6.122889026, 6.125740703, 6.125890218, 6.134166495, 6.135570908, 6.139387846, 6.140574231, 6.141574437, 6.14442153, 6.146428244, 6.150324809, 6.15065191, 6.1519437, 6.152584153, 6.153254394, 6.154379769, 6.159462432, 6.160332027, 6.160558878, 6.161467657, 6.161821452, 6.163750258, 6.166672166, 6.16752681, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.170062556, 6.171541097, 6.172124837, 6.174530831, 6.176230838, 6.177207213, 6.178999106, 6.18079415, 6.180987489, 6.183202149, 6.184249673, 6.185046745, 6.188128136, 6.188133292, 6.189258094, 6.189763067, 6.190169794, 6.190578814, 6.193393827, 6.193959236, 6.194198976, 6.19727693, 6.200274971, 6.204112102, 6.205722829, 6.20850519, 6.20850519, 6.209636007, 6.210690635, 6.212935657, 6.213274215, 6.213699846, 6.216149378, 6.216469031, 6.219832268, 6.223137646, 6.223249467, 6.223283151, 6.225970991, 6.232015192, 6.232618409, 6.233183245, 6.238464116, 6.239264109, 6.24015089, 6.241044545, 6.244015948, 6.246131505, 6.248309789, 6.255495979, 6.256873124, 6.256885727, 6.258532687, 6.259006439, 6.260959304, 6.262000756, 6.264780109, 6.268044385, 6.270169107, 6.271061044, 6.273784973, 6.275540203, 6.276225768, 6.276408366, 6.276785162, 6.279219083, 6.280081804, 6.282279852, 6.287830824, 6.287863477, 6.290839463, 6.291271396, 6.2916111, 6.292562613, 6.29300629, 6.29345054, 6.29345054, 6.29345054, 6.29345054, 6.29345054, 6.29345054, 6.29345054, 6.293644738, 6.298086375, 6.29864405, 6.299046883, 6.30049151, 6.303602547, 6.303629614, 6.306652574, 6.309342132, 6.313452946, 6.316212965, 6.318026053, 6.319198116, 6.319527394, 6.321909157, 6.322649687, 6.325152952, 6.328272152, 6.328385005, 6.329338523, 6.330421221, 6.331449211, 6.331618776, 6.333665017, 6.335899154, 6.336237712, 6.339782258, 6.340897178, 6.341008455, 6.342444318, 6.343218533, 6.343703742, 6.34515364, 6.34568525, 6.3474084, 6.347776174, 6.350324809, 6.35238709, 6.353230047, 6.354601464, 6.356541211, 6.357160411, 6.360701519, 6.362198391, 6.36877449, 6.371908297, 6.375152952, 6.375701176, 6.375830068, 6.381680071, 6.382107995, 6.383117367, 6.383574792, 6.385809789, 6.386760586, 6.389446278, 6.38983897, 6.391535483, 6.393276105, 6.394886403, 6.395990582, 6.402923283, 6.405025664, 6.408674469, 6.409144497, 6.411234275, 6.411863958, 6.413047364, 6.415169794, 6.416135629, 6.416838523, 6.416838523, 6.416838523, 6.416838523, 6.418345277, 6.418382943, 6.41893002, 6.419072661, 6.419072661, 6.419072661, 6.419410646, 6.420032767, 6.421107015, 6.421731285, 6.423125902, 6.427187307, 6.431185124, 6.434137623, 6.437794734, 6.443911975, 6.445524278, 6.447683371, 6.448661809, 6.450171284, 6.450174435, 6.450689146, 6.451012236, 6.451228776, 6.453517907, 6.45598663, 6.461010861, 6.461748127, 6.46228432, 6.462708806, 6.465828865, 6.466237884, 6.46697429, 6.467969341, 6.468898226, 6.469052898, 6.469244518, 6.471399601, 6.472068697, 6.476649825, 6.477933595, 6.478181068, 6.4851722, 6.486855537, 6.487481955, 6.488010701, 6.490989551, 6.493354873, 6.496962862, 6.497918815, 6.49804513, 6.498540936, 6.498540936, 6.501094727, 6.506638253, 6.507090809, 6.508136557, 6.509936327, 6.509983158, 6.512264556, 6.512591943, 6.519488125, 6.52067322, 6.520923558, 6.522437616, 6.525594624, 6.526669302, 6.527028769, 6.52733897, 6.530421507, 6.530913018, 6.531083729, 6.531590133, 6.531967072, 6.53309789, 6.536227974, 6.537933938, 6.53799237, 6.540226507, 6.540226507, 6.540226507, 6.540226507, 6.540226507, 6.540226507, 6.540226507, 6.540226507, 6.54318588, 6.544019385, 6.544146559, 6.544675878, 6.548365642, 6.549163058, 6.551843164, 6.552496219, 6.552650031, 6.553866777, 6.556269334, 6.556959625, 6.557064171, 6.558292088, 6.559789819, 6.559964254, 6.559964254, 6.56070023, 6.565083293, 6.565899326, 6.567460645, 6.569127655, 6.570260191, 6.572390069, 6.573114732, 6.57359593, 6.575826631, 6.575933755, 6.57993057, 6.581594143, 6.582033237, 6.583164914, 6.583496597, 6.585813799, 6.586557366, 6.587586788, 6.588990857, 6.592837011, 6.597973809, 6.598455008, 6.598793566, 6.605261681, 6.605355056, 6.606590706, 6.607649344, 6.608352238, 6.611115694, 6.612142824, 6.613132433, 6.614463463, 6.614518457, 6.616577301, 6.62019374, 6.62057297, 6.621676291, 6.622181549, 6.624963051, 6.626716849, 6.626738846, 6.627301162, 6.628136386, 6.630760638, 6.631740221, 6.63433354, 6.636250315, 6.636426755, 6.637935657, 6.641872465, 6.644392315, 6.64508948, 6.650867877, 6.652353578, 6.659858992, 6.661945762, 6.663614491, 6.663614491, 6.663614491, 6.664536789, 6.668082766, 6.66874155, 6.668911116, 6.669145987, 6.671502715, 6.674661442, 6.675396989, 6.677019317, 6.678328006, 6.680214133, 6.685586375, 6.685936963, 6.686747554, 6.690848629, 6.692940125, 6.697587131, 6.701962604, 6.703447446, 6.70591359, 6.71049701, 6.724415687, 6.724678198, 6.726914484, 6.729110813, 6.731487592, 6.733518939, 6.737760363, 6.738226954, 6.744947412, 6.747719891, 6.753456039, 6.755977177, 6.762198391, 6.7660265, 6.769392315, 6.769947412, 6.770565237, 6.771195092, 6.771483811, 6.774058225, 6.77565993, 6.778724995, 6.782730116, 6.783362549, 6.784599918, 6.785222039, 6.785779714, 6.785779714, 6.786331374, 6.786521276, 6.787983089, 6.788226668, 6.789625868, 6.790997571, 6.793278683, 6.794236784, 6.794583076, 6.797803671, 6.797856087, 6.798736853, 6.80163092, 6.802894068, 6.803016945, 6.804911666, 6.81206864, 6.81347185, 6.814094544, 6.814372379, 6.816101258, 6.818118856, 6.820939025, 6.82138127, 6.824783461, 6.828016945, 6.83825531, 6.842147866, 6.844434419, 6.84450574, 6.845606053, 6.846531931, 6.847606036, 6.849764556, 6.856113804, 6.856283942, 6.856580395, 6.862891833, 6.866945934, 6.869320994, 6.870827318, 6.871574036, 6.874143294, 6.874360693, 6.879416718, 6.885445453, 6.887255104, 6.887629752, 6.892144428, 6.892345501, 6.897465113, 6.91038874, 6.910390459, 6.911148347, 6.923916443, 6.925886781, 6.926274031, 6.928262414, 6.930223586, 6.93090242, 6.93216127, 6.933041177, 6.93382914, 6.935955867, 6.938013852, 6.940479137, 6.945890046, 6.948738572, 6.949047914, 6.950075617, 6.959947068, 6.962548979, 6.963188286, 6.967568055, 6.974840173, 6.975037808, 6.976029422, 6.97916323, 6.991365058, 6.992092871, 6.993280401, 6.994835705, 6.995485323, 6.997267478, 7.00263972, 7.003399326, 7.003940675, 7.006375885, 7.006599299, 7.009409157, 7.013188286, 7.016101258, 7.017345501, 7.02565993, 7.025924589, 7.027668935, 7.029517541, 7.03090156, 7.032730116, 7.03601258, 7.037675294, 7.04155496, 7.042053917, 7.044929367, 7.04801219, 7.048425792, 7.049623061, 7.050853269, 7.057675122, 7.06278786, 7.064965285, 7.06531931, 7.066860005, 7.067345501, 7.067734327, 7.068278683, 7.07892263, 7.080012202, 7.086689695, 7.088103504, 7.088217502, 7.088246718, 7.088528563, 7.088669485, 7.089631883, 7.091780092, 7.092411666, 7.093063003, 7.095462982, 7.101947137, 7.104275796, 7.110935244, 7.112143397, 7.112341033, 7.114121468, 7.120481027, 7.13524839, 7.141044545, 7.145939598, 7.148365642, 7.149953599, 7.151802777, 7.153507596, 7.156288238, 7.156599299, 7.160842442, 7.161850095, 7.162229326, 7.165594281, 7.168280401, 7.171470063, 7.17209734, 7.175037808, 7.177253901, 7.184026775, 7.18703169, 7.1933543, 7.201984945, 7.210002922, 7.212198391, 7.225689146, 7.226135973, 7.227328659, 7.245155931, 7.246813776, 7.251135973, 7.251838294, 7.258211315, 7.266210387, 7.269909603, 7.273737283, 7.281740221, 7.281878566, 7.28420121, 7.287024816, 7.287198391, 7.288896336, 7.299444903, 7.306736784, 7.312509452, 7.312674435, 7.314461745, 7.316723379, 7.321542815, 7.325970991, 7.331665177, 7.331685227, 7.337887537, 7.338499347, 7.338841342, 7.341280562, 7.343599196, 7.344890699, 7.349133842, 7.349284217, 7.353545405, 7.35503884, 7.359766619, 7.361265209, 7.378883962, 7.379195023, 7.379506084, 7.383790902, 7.384964254, 7.387822231, 7.395975115, 7.398906991, 7.400037808, 7.401914484, 7.406719599, 7.406740221, 7.409031645, 7.411826322, 7.41338764, 7.417468378, 7.421993366, 7.42565993, 7.431796934, 7.46038874, 7.460586375, 7.46084416, 7.465989551, 7.469830549, 7.472124837, 7.476786451, 7.479715749, 7.485077679, 7.491780092, 7.495660617, 7.50065993, 7.503883962, 7.515733485, 7.516592768, 7.517967622, 7.520249307, 7.522324191, 7.524415687, 7.529817144, 7.533749227, 7.534880044, 7.536010861, 7.538721042, 7.541931326, 7.545994019, 7.546813776, 7.547124837, 7.55674194, 7.557164707, 7.557824637, 7.568648175, 7.575513852, 7.586265209, 7.58881986, 7.592713274, 7.592912628, 7.597162645, 7.598114732, 7.605016498, 7.605695332, 7.614121468, 7.618278683, 7.621823228, 7.625970991, 7.631260741, 7.633108201, 7.63736853, 7.637877226, 7.639753901, 7.640793634, 7.647737792, 7.647803671, 7.648114732, 7.664597512, 7.6738245, 7.685823537, 7.700716643, 7.707109713, 7.71355606, 7.724047914, 7.744405204, 7.749047914, 7.754648725, 7.756740221, 7.757132055, 7.765733485, 7.769100158, 7.781006393, 7.782164707, 7.782419055, 7.789753042, 7.81038874, 7.823568433, 7.850802571, 7.861401835, 7.862198391, 7.865254004, 7.867199423, 7.868703169, 7.8744724, 7.885077679, 7.894560734, 7.933352238, 7.955123909, 7.955495979, 7.958729463, 7.961883034, 7.962198391, 7.969184368, 7.988499347, 7.99601722, 8.042967622, 8.087566165, 8.08952705, 8.094702516, 8.114121468, 8.124047914, 8.126141129, 8.175943494, 8.177612222, 8.185275314, 8.189432529, 8.189432529, 8.190733485, 8.202894068, 8.208041177, 8.210586375, 8.250970991, 8.343278683, 8.368278683, 8.371588644, 8.384964254, 8.385963601, 8.390479137, 8.394608854, 8.396502715, 8.400348869, 8.418278683, 8.425970991, 8.436953496, 8.44071458, 8.44749261, 8.452894068, 8.459964254, 8.466355606, 8.482193923, 8.508352238, 8.52565993, 8.564121468, 8.571672854, 8.581740221, 8.585586375, 8.593957517, 8.637509452, 8.641355606, 8.661887331, 8.662198391, 8.681429161, 8.696813776, 8.716723379, 8.721191655, 8.722294975, 8.78340895, 8.842345501, 8.859342132, 8.889121468, 8.955184918, 8.99749261, 9.036321922, 9.05295078, 9.274585825, 9.325970991, 9.379817144, 9.397124837];
